Start refactoring the notifications
