+++ identity-service/src/notifications/index.js
@@ -9,6 +9,11 @@ const uuidv4 = require('uuid/v4')
 const getEmailNotifications = require('./fetchNotificationMetadata')
 const renderEmail = require('./renderEmail')
 const { logger } = require('../logging')
+const {
+  updateBlockchainIds,
+  calculateTrackListenMilestones,
+  getHighestBlockNumber 
+} = require('./utils')
 
 const notificationTypes = Object.freeze({
   Follow: 'Follow',
@@ -59,8 +64,6 @@ const repostMilestoneList = baseMilestoneList
 const favoriteMilestoneList = baseMilestoneList
 // Track listen milestone list
 const trackListenMilestoneList = baseMilestoneList
-// Number of tracks to fetch for new listens on each poll
-const trackListenMilestonePollCount = 100
 
 const notifDiscProv = config.get('notificationDiscoveryProvider')
 const emailCachePath = './emailCache'
@@ -77,10 +80,21 @@ class NotificationProcessor {
       { redis:
         { port: config.get('redisPort'), host: config.get('redisHost') }
       })
-    this.startBlock = config.get('notificationStartBlock')
     this.blockchainIdsPopulated = false
   }
 
+  /**
+   * Steps in init flow
+   * 1. Clear the notifQueue and emailQueue
+   * 2. Update all blockchainId's in the users table where blockchainId is null
+   * 3. Process notif queue
+   *    1....
+   * 4. Process email queue
+   *    1....
+   * @param {Object} audiusLibs libs instance
+   * @param {Object} expressApp express app context
+   * @param {Object} redis redis connection
+   */
   async init (audiusLibs, expressApp, redis) {
     // Clear any pending notif jobs
     await this.notifQueue.empty()
@@ -92,7 +106,7 @@ class NotificationProcessor {
     this.mg = this.expressApp.get('mailgun')
 
     // Index all blockchain ids
-    await this.updateBlockchainIds()
+    await updateBlockchainIds()
 
     // Notification processing job
     // Indexes network notifications
@@ -145,61 +159,12 @@ class NotificationProcessor {
       { repeat: { cron: '0 * * * *' } }
     )
 
-    let startBlock = await this.getHighestBlockNumber()
+    let startBlock = await getHighestBlockNumber()
     logger.info(`Starting with ${startBlock}`)
     await this.notifQueue.add({
       minBlock: startBlock,
       type: notificationJobType
     })
-  }
-
-  async getHighestBlockNumber () {
-    let highestBlockNumber = await models.NotificationAction.max('blocknumber')
-    if (!highestBlockNumber) {
-      highestBlockNumber = this.startBlock
-    }
-    let date = new Date()
-    logger.info(`Highest block: ${highestBlockNumber} - ${date}`)
-    return highestBlockNumber
-  }
-
-  async calculateTrackListenMilestones () {
-    // Select last 10 distinct tracks with listens
-    let recentListenCountQuery = {
-      attributes: [[models.Sequelize.col('trackId'), 'trackId'], [models.Sequelize.fn('max', models.Sequelize.col('hour')), 'hour']],
-      order: [[models.Sequelize.col('hour'), 'DESC']],
-      group: ['trackId'],
-      limit: trackListenMilestonePollCount
-    }
-
-    // Distinct tracks
-    let res = await models.TrackListenCount.findAll(recentListenCountQuery)
-    let tracksListenedTo = res.map((listenEntry) => listenEntry.trackId)
-
-    // Total listens query
-    let totalListens = {
-      attributes: [
-        [models.Sequelize.col('trackId'), 'trackId'],
-        [
-          models.Sequelize.fn('date_trunc', 'millennium', models.Sequelize.col('hour')),
-          'date'
-        ],
-        [models.Sequelize.fn('sum', models.Sequelize.col('listens')), 'listens']
-      ],
-      group: ['trackId', 'date'],
-      order: [[models.Sequelize.col('listens'), 'DESC']],
-      where: {
-        trackId: { [models.Sequelize.Op.in]: tracksListenedTo }
-      }
-    }
-
-    // Map of listens
-    let totalListenQuery = await models.TrackListenCount.findAll(totalListens)
-    let processedTotalListens = totalListenQuery.map((x) => {
-      return { trackId: x.trackId, listenCount: x.listens }
-    })
-
-    return processedTotalListens
   }
 
   async indexMilestones (milestones, owners, metadata, listenCounts, tx) {
@@ -532,11 +497,16 @@ class NotificationProcessor {
     }
   }
 
+  /**
+   * 1. Get the total listens for the most reecently listened to tracks
+   * 2. Query the discprov for new notifications starting at minBlock
+   * @param {Integer} minBlock min start block to start querying discprov for new notifications
+   */
   async indexNotifications (minBlock) {
     logger.info(`${new Date()} - indexNotifications job`)
 
     // Query owners for tracks relevant to track listen counts
-    let listenCounts = await this.calculateTrackListenMilestones()
+    let listenCounts = await calculateTrackListenMilestones()
     let trackIdOwnersToRequestList = listenCounts.map(x => x.trackId)
 
     // These track_id get parameters will be used to retrieve track owner info
@@ -569,364 +539,23 @@ class NotificationProcessor {
 
         // Handle the 'follow' notification type
         if (notif.type === notificationTypes.Follow) {
-          let notificationTarget = notif.metadata.followee_user_id
-          // Skip notification based on user settings
-          let userNotifSettings = await models.UserNotificationSettings.findOne(
-            { where: { userId: notificationTarget }, transaction: tx }
-          )
-          if (userNotifSettings && !userNotifSettings.followers) {
-            continue
-          }
-
-          let notificationInitiator = notif.metadata.follower_user_id
-          let unreadQuery = await models.Notification.findAll({
-            where: {
-              isViewed: false,
-              userId: notificationTarget,
-              type: notificationTypes.Follow
-            },
-            transaction: tx
-          })
-
-          let notificationId = null
-          // Insertion into the Notification table
-          // Follow - userId = notif target, entityId=null, actionEntityId = user who followed target
-          if (unreadQuery.length === 0) {
-            let createNotifTx = await models.Notification.create({
-              type: notificationTypes.Follow,
-              isViewed: false,
-              isRead: false,
-              isHidden: false,
-              userId: notificationTarget,
-              blocknumber: blocknumber,
-              timestamp: timestamp
-            }, { transaction: tx })
-            notificationId = createNotifTx.id
-          } else {
-            notificationId = unreadQuery[0].id
-          }
-
-          if (notificationId) {
-            // Insertion into the NotificationActions table
-            let notifActionCreateTx = await models.NotificationAction.findOrCreate({
-              where: {
-                notificationId: notificationId,
-                actionEntityType: actionEntityTypes.User,
-                actionEntityId: notificationInitiator,
-                blocknumber
-              },
-              transaction: tx
-            })
-            // TODO: Handle log statements to indicate how many notifs have been processed
-            let updatePerformed = notifActionCreateTx[1]
-            if (updatePerformed) {
-              // Update Notification table timestamp
-              let newNotificationTimestamp = notifActionCreateTx[0].createdAt
-              await models.Notification.update({
-                timestamp: newNotificationTimestamp
-              }, {
-                where: { id: notificationId },
-                returning: true,
-                plain: true,
-                transaction: tx
-              })
-            }
-          }
+          await _processFollowNotifications(notif, blocknumber, timestamp, tx)
         }
 
         // Handle the 'repost' notification type
         // track/album/playlist
         if (notif.type === notificationTypes.Repost.base) {
-          let repostType = null
-          switch (notif.metadata.entity_type) {
-            case 'track':
-              repostType = notificationTypes.Repost.track
-              break
-            case 'album':
-              repostType = notificationTypes.Repost.album
-              break
-            case 'playlist':
-              repostType = notificationTypes.Repost.playlist
-              break
-            default:
-              throw new Error('Invalid repost type')
-          }
-          let notificationTarget = notif.metadata.entity_owner_id
-
-          // Skip notification based on user settings
-          let userNotifSettings = await models.UserNotificationSettings.findOne(
-            { where: { userId: notificationTarget }, transaction: tx }
-          )
-          if (userNotifSettings && !userNotifSettings.reposts) {
-            continue
-          }
-
-          let notificationEntityId = notif.metadata.entity_id
-          let notificationInitiator = notif.initiator
-
-          let unreadQuery = await models.Notification.findAll({
-            where: {
-              isViewed: false,
-              userId: notificationTarget,
-              type: repostType,
-              entityId: notificationEntityId
-            },
-            transaction: tx
-          })
-
-          let notificationId = null
-          // Insert new notification
-          // Repost - userId=notif target, entityId=track/album/repost id, actionEntityType=User actionEntityId=user who reposted
-          // As multiple users repost an entity, NotificationActions are added matching the NotificationId
-          if (unreadQuery.length === 0) {
-            let repostNotifTx = await models.Notification.create({
-              type: repostType,
-              isRead: false,
-              isHidden: false,
-              isViewed: false,
-              userId: notificationTarget,
-              entityId: notificationEntityId,
-              blocknumber,
-              timestamp
-            }, { transaction: tx })
-            notificationId = repostNotifTx.id
-          } else {
-            notificationId = unreadQuery[0].id
-          }
-
-          if (notificationId) {
-            let notifActionCreateTx = await models.NotificationAction.findOrCreate({
-              where: {
-                notificationId: notificationId,
-                actionEntityType: actionEntityTypes.User,
-                actionEntityId: notificationInitiator,
-                blocknumber
-              },
-              transaction: tx
-            })
-            // Update Notification table timestamp
-            let updatePerformed = notifActionCreateTx[1]
-            if (updatePerformed) {
-              let newNotificationTimestamp = notifActionCreateTx[0].createdAt
-              await models.Notification.update({
-                timestamp: newNotificationTimestamp
-              }, {
-                where: { id: notificationId },
-                returning: true,
-                plain: true,
-                transaction: tx
-              })
-            }
-          }
+          await _processBaseRepostNotifications(notif, blocknumber, timestamp, tx)
         }
 
         // Handle the 'favorite' notification type, track/album/playlist
         if (notif.type === notificationTypes.Favorite.base) {
-          let favoriteType = null
-          switch (notif.metadata.entity_type) {
-            case 'track':
-              favoriteType = notificationTypes.Favorite.track
-              break
-            case 'album':
-              favoriteType = notificationTypes.Favorite.album
-              break
-            case 'playlist':
-              favoriteType = notificationTypes.Favorite.playlist
-              break
-            default:
-              throw new Error('Invalid favorite type')
-          }
-          let notificationTarget = notif.metadata.entity_owner_id
-          // Skip notification based on user settings
-          let userNotifSettings = await models.UserNotificationSettings.findOne(
-            { where: { userId: notificationTarget }, transaction: tx }
-          )
-          if (userNotifSettings && !userNotifSettings.favorites) {
-            continue
-          }
-
-          let notificationEntityId = notif.metadata.entity_id
-          let notificationInitiator = notif.initiator
-          let unreadQuery = await models.Notification.findAll({
-            where: {
-              isViewed: false,
-              userId: notificationTarget,
-              type: favoriteType,
-              entityId: notificationEntityId
-            },
-            transaction: tx
-          })
-
-          let notificationId = null
-          if (unreadQuery.length === 0) {
-          // Favorite - userId=notif target, entityId=track/album/repost id, actionEntityType=User actionEntityId=user who favorited
-          // As multiple users favorite an entity, NotificationActions are added matching the NotificationId
-            let favoriteNotifTx = await models.Notification.create({
-              type: favoriteType,
-              isViewed: false,
-              isRead: false,
-              isHidden: false,
-              userId: notificationTarget,
-              entityId: notificationEntityId,
-              blocknumber,
-              timestamp
-            }, { transaction: tx })
-            notificationId = favoriteNotifTx.id
-          } else {
-            notificationId = unreadQuery[0].id
-          }
-
-          if (notificationId) {
-            let notifActionCreateTx = await models.NotificationAction.findOrCreate({
-              where: {
-                notificationId: notificationId,
-                actionEntityType: actionEntityTypes.User,
-                actionEntityId: notificationInitiator,
-                blocknumber
-              },
-              transaction: tx
-            })
-            // Update Notification table timestamp
-            let updatePerformed = notifActionCreateTx[1]
-            if (updatePerformed) {
-              let newNotificationTimestamp = notifActionCreateTx[0].createdAt
-              await models.Notification.update({
-                timestamp: newNotificationTimestamp
-              }, {
-                where: { id: notificationId },
-                returning: true,
-                plain: true,
-                transaction: tx
-              })
-            }
-          }
+          await _processFavoriteNotifications(notif, blocknumber, timestamp, tx)
         }
 
         // Handle the 'create' notification type, track/album/playlist
         if (notif.type === notificationTypes.Create.base) {
-          let createType = null
-          let actionEntityType = null
-          switch (notif.metadata.entity_type) {
-            case 'track':
-              createType = notificationTypes.Create.track
-              actionEntityType = actionEntityTypes.Track
-              break
-            case 'album':
-              createType = notificationTypes.Create.album
-              actionEntityType = actionEntityTypes.User
-              break
-            case 'playlist':
-              createType = notificationTypes.Create.playlist
-              actionEntityType = actionEntityTypes.User
-              break
-            default:
-              throw new Error('Invalid create type')
-          }
-
-          // Query user IDs from subscriptions table
-          // Notifications go to all users subscribing to this track uploader
-          let subscribers = await models.Subscription.findAll({
-            where: {
-              userId: notif.initiator
-            },
-            transaction: tx
-          })
-
-          // No operation if no users subscribe to this creator
-          if (subscribers.length === 0) { continue }
-
-          // The notification entity id is the uploader id for tracks
-          // Each track will added to the notification actions table
-          // For playlist/albums, the notification entity id is the collection id itself
-          let notificationEntityId =
-            actionEntityType === actionEntityTypes.Track
-              ? notif.initiator
-              : notif.metadata.entity_id
-
-          // Action table entity is trackId for CreateTrack notifications
-          // Allowing multiple track creates to be associated w/ a single notif for your subscription
-          // For collections, the entity is the owner id, producing a distinct notif for each
-          let createdActionEntityId =
-            actionEntityType === actionEntityTypes.Track
-              ? notif.metadata.entity_id
-              : notif.metadata.entity_owner_id
-
-          // Create notification for each subscriber
-          await Promise.all(subscribers.map(async (s) => {
-            // Add notification for this user indicating the uploader has added a track
-            let notificationTarget = s.subscriberId
-
-            let unreadQuery = await models.Notification.findAll({
-              where: {
-                isViewed: false,
-                userId: notificationTarget,
-                type: createType,
-                entityId: notificationEntityId
-              },
-              transaction: tx
-            })
-
-            let notificationId = null
-            if (unreadQuery.length === 0) {
-              let createTrackNotifTx = await models.Notification.create({
-                isViewed: false,
-                isRead: false,
-                isHidden: false,
-                userId: notificationTarget,
-                type: createType,
-                entityId: notificationEntityId,
-                blocknumber,
-                timestamp
-              }, { transaction: tx })
-              notificationId = createTrackNotifTx.id
-            } else {
-              notificationId = unreadQuery[0].id
-            }
-
-            if (notificationId) {
-              // Action entity id can be one of album/playlist/track
-              let notifActionCreateTx = await models.NotificationAction.findOrCreate({
-                where: {
-                  notificationId,
-                  actionEntityType: actionEntityType,
-                  actionEntityId: createdActionEntityId,
-                  blocknumber
-                },
-                transaction: tx
-              })
-
-              // Update Notification table timestamp
-              let updatePerformed = notifActionCreateTx[1]
-              if (updatePerformed) {
-                let newNotificationTimestamp = notifActionCreateTx[0].createdAt
-                await models.Notification.update({
-                  timestamp: newNotificationTimestamp
-                }, {
-                  where: { id: notificationId },
-                  returning: true,
-                  plain: true,
-                  transaction: tx
-                })
-              }
-            }
-          }))
-
-          // Dedupe album /playlist notification
-          if (createType === notificationTypes.Create.album ||
-              createType === notificationTypes.Create.playlist) {
-            let trackIdList = notif.metadata.collection_content.track_ids
-            if (trackIdList.length > 0) {
-              for (var entry of trackIdList) {
-                let trackId = entry.track
-                await models.NotificationAction.destroy({
-                  where: {
-                    actionEntityType: actionEntityTypes.Track,
-                    actionEntityId: trackId
-                  }
-                }, { transaction: tx })
-              }
-            }
-          }
+          await _processCreateNotifications(notif, blocknumber, timestamp, tx)
         }
       }
 
@@ -949,57 +578,6 @@ class NotificationProcessor {
       await tx.rollback()
     }
     return metadata.max_block_number
-  }
-
-  async updateBlockchainIds () {
-    // For any users missing blockchain id, here we query from the disc prov and fill in values
-    let usersWithoutBlockchainId = await models.User.findAll({
-      attributes: ['walletAddress', 'handle'],
-      where: { blockchainUserId: null }
-    })
-    for (let updateUser of usersWithoutBlockchainId) {
-      try {
-        let walletAddress = updateUser.walletAddress
-        logger.info(`Updating user with wallet ${walletAddress}`)
-        const response = await axios({
-          method: 'get',
-          url: `${notifDiscProv}/users`,
-          params: {
-            wallet: walletAddress
-          }
-        })
-        if (response.data.data.length === 1) {
-          let respUser = response.data.data[0]
-          let missingUserId = respUser.user_id
-          let missingHandle = respUser.handle
-          let updateObject = { blockchainUserId: missingUserId }
-
-          if (updateUser.handle === null) {
-            updateObject['handle'] = missingHandle
-          }
-          await models.User.update(
-            updateObject,
-            { where: { walletAddress } }
-          )
-          logger.info(`Updated wallet ${walletAddress} to blockchainUserId: ${missingUserId}, ${updateUser.handle}`)
-          continue
-        }
-        for (let respUser of response.data.data) {
-          // Only update if handles match
-          if (respUser.handle === updateUser.handle) {
-            let missingUserId = respUser.user_id
-            await models.User.update(
-              { blockchainUserId: missingUserId },
-              { where: { walletAddress, handle: updateUser.handle } }
-            )
-            logger.info(`Updated wallet ${walletAddress} to blockchainUserId: ${missingUserId}, ${updateUser.handle}`)
-            await models.UserNotificationSettings.findOrCreate({ where: { userId: missingUserId } })
-          }
-        }
-      } catch (e) {
-        logger.error(e)
-      }
-    }
   }
 
   async processEmailNotifications () {
@@ -1311,3 +889,361 @@ class NotificationProcessor {
 }
 
 module.exports = NotificationProcessor
+
+
+async function _processFollowNotifications(notif, blocknumber, timestamp, tx) {
+  let notificationTarget = notif.metadata.followee_user_id
+  // Skip notification based on user settings
+  let userNotifSettings = await models.UserNotificationSettings.findOne(
+    { where: { userId: notificationTarget }, transaction: tx }
+  )
+  if (userNotifSettings && !userNotifSettings.followers) {
+    return
+  }
+
+  let notificationInitiator = notif.metadata.follower_user_id
+  let unreadQuery = await models.Notification.findAll({
+    where: {
+      isViewed: false,
+      userId: notificationTarget,
+      type: notificationTypes.Follow
+    },
+    transaction: tx
+  })
+
+  let notificationId = null
+  // Insertion into the Notification table
+  // Follow - userId = notif target, entityId=null, actionEntityId = user who followed target
+  if (unreadQuery.length === 0) {
+    let createNotifTx = await models.Notification.create({
+      type: notificationTypes.Follow,
+      isViewed: false,
+      isRead: false,
+      isHidden: false,
+      userId: notificationTarget,
+      blocknumber: blocknumber,
+      timestamp: timestamp
+    }, { transaction: tx })
+    notificationId = createNotifTx.id
+  } else {
+    notificationId = unreadQuery[0].id
+  }
+
+  if (notificationId) {
+    // Insertion into the NotificationActions table
+    let notifActionCreateTx = await models.NotificationAction.findOrCreate({
+      where: {
+        notificationId: notificationId,
+        actionEntityType: actionEntityTypes.User,
+        actionEntityId: notificationInitiator,
+        blocknumber
+      },
+      transaction: tx
+    })
+    // TODO: Handle log statements to indicate how many notifs have been processed
+    let updatePerformed = notifActionCreateTx[1]
+    if (updatePerformed) {
+      // Update Notification table timestamp
+      let newNotificationTimestamp = notifActionCreateTx[0].createdAt
+      await models.Notification.update({
+        timestamp: newNotificationTimestamp
+      }, {
+        where: { id: notificationId },
+        returning: true,
+        plain: true,
+        transaction: tx
+      })
+    }
+  }
+}
+
+async function _processBaseRepostNotifications(notif, blocknumber, timestamp, tx) {
+  let repostType = null
+  switch (notif.metadata.entity_type) {
+    case 'track':
+      repostType = notificationTypes.Repost.track
+      break
+    case 'album':
+      repostType = notificationTypes.Repost.album
+      break
+    case 'playlist':
+      repostType = notificationTypes.Repost.playlist
+      break
+    default:
+      throw new Error('Invalid repost type')
+  }
+  let notificationTarget = notif.metadata.entity_owner_id
+
+  // Skip notification based on user settings
+  let userNotifSettings = await models.UserNotificationSettings.findOne(
+    { where: { userId: notificationTarget }, transaction: tx }
+  )
+  if (userNotifSettings && !userNotifSettings.reposts) {
+    return
+  }
+
+  let notificationEntityId = notif.metadata.entity_id
+  let notificationInitiator = notif.initiator
+
+  let unreadQuery = await models.Notification.findAll({
+    where: {
+      isViewed: false,
+      userId: notificationTarget,
+      type: repostType,
+      entityId: notificationEntityId
+    },
+    transaction: tx
+  })
+
+  let notificationId = null
+  // Insert new notification
+  // Repost - userId=notif target, entityId=track/album/repost id, actionEntityType=User actionEntityId=user who reposted
+  // As multiple users repost an entity, NotificationActions are added matching the NotificationId
+  if (unreadQuery.length === 0) {
+    let repostNotifTx = await models.Notification.create({
+      type: repostType,
+      isRead: false,
+      isHidden: false,
+      isViewed: false,
+      userId: notificationTarget,
+      entityId: notificationEntityId,
+      blocknumber,
+      timestamp
+    }, { transaction: tx })
+    notificationId = repostNotifTx.id
+  } else {
+    notificationId = unreadQuery[0].id
+  }
+
+  if (notificationId) {
+    let notifActionCreateTx = await models.NotificationAction.findOrCreate({
+      where: {
+        notificationId: notificationId,
+        actionEntityType: actionEntityTypes.User,
+        actionEntityId: notificationInitiator,
+        blocknumber
+      },
+      transaction: tx
+    })
+    // Update Notification table timestamp
+    let updatePerformed = notifActionCreateTx[1]
+    if (updatePerformed) {
+      let newNotificationTimestamp = notifActionCreateTx[0].createdAt
+      await models.Notification.update({
+        timestamp: newNotificationTimestamp
+      }, {
+        where: { id: notificationId },
+        returning: true,
+        plain: true,
+        transaction: tx
+      })
+    }
+  }
+}
+
+async function _processFavoriteNotifications(notif, blocknumber, timestamp, tx) {
+  let favoriteType = null
+  switch (notif.metadata.entity_type) {
+    case 'track':
+      favoriteType = notificationTypes.Favorite.track
+      break
+    case 'album':
+      favoriteType = notificationTypes.Favorite.album
+      break
+    case 'playlist':
+      favoriteType = notificationTypes.Favorite.playlist
+      break
+    default:
+      throw new Error('Invalid favorite type')
+  }
+  let notificationTarget = notif.metadata.entity_owner_id
+  // Skip notification based on user settings
+  let userNotifSettings = await models.UserNotificationSettings.findOne(
+    { where: { userId: notificationTarget }, transaction: tx }
+  )
+  if (userNotifSettings && !userNotifSettings.favorites) {
+    return
+  }
+
+  let notificationEntityId = notif.metadata.entity_id
+  let notificationInitiator = notif.initiator
+  let unreadQuery = await models.Notification.findAll({
+    where: {
+      isViewed: false,
+      userId: notificationTarget,
+      type: favoriteType,
+      entityId: notificationEntityId
+    },
+    transaction: tx
+  })
+
+  let notificationId = null
+  if (unreadQuery.length === 0) {
+  // Favorite - userId=notif target, entityId=track/album/repost id, actionEntityType=User actionEntityId=user who favorited
+  // As multiple users favorite an entity, NotificationActions are added matching the NotificationId
+    let favoriteNotifTx = await models.Notification.create({
+      type: favoriteType,
+      isViewed: false,
+      isRead: false,
+      isHidden: false,
+      userId: notificationTarget,
+      entityId: notificationEntityId,
+      blocknumber,
+      timestamp
+    }, { transaction: tx })
+    notificationId = favoriteNotifTx.id
+  } else {
+    notificationId = unreadQuery[0].id
+  }
+
+  if (notificationId) {
+    let notifActionCreateTx = await models.NotificationAction.findOrCreate({
+      where: {
+        notificationId: notificationId,
+        actionEntityType: actionEntityTypes.User,
+        actionEntityId: notificationInitiator,
+        blocknumber
+      },
+      transaction: tx
+    })
+    // Update Notification table timestamp
+    let updatePerformed = notifActionCreateTx[1]
+    if (updatePerformed) {
+      let newNotificationTimestamp = notifActionCreateTx[0].createdAt
+      await models.Notification.update({
+        timestamp: newNotificationTimestamp
+      }, {
+        where: { id: notificationId },
+        returning: true,
+        plain: true,
+        transaction: tx
+      })
+    }
+  }
+}
+
+async function _processCreateNotifications(notif, blocknumber, timestamp, tx) {
+  let createType = null
+  let actionEntityType = null
+  switch (notif.metadata.entity_type) {
+    case 'track':
+      createType = notificationTypes.Create.track
+      actionEntityType = actionEntityTypes.Track
+      break
+    case 'album':
+      createType = notificationTypes.Create.album
+      actionEntityType = actionEntityTypes.User
+      break
+    case 'playlist':
+      createType = notificationTypes.Create.playlist
+      actionEntityType = actionEntityTypes.User
+      break
+    default:
+      throw new Error('Invalid create type')
+  }
+
+  // Query user IDs from subscriptions table
+  // Notifications go to all users subscribing to this track uploader
+  let subscribers = await models.Subscription.findAll({
+    where: {
+      userId: notif.initiator
+    },
+    transaction: tx
+  })
+
+  // No operation if no users subscribe to this creator
+  if (subscribers.length === 0) { return }
+
+  // The notification entity id is the uploader id for tracks
+  // Each track will added to the notification actions table
+  // For playlist/albums, the notification entity id is the collection id itself
+  let notificationEntityId =
+    actionEntityType === actionEntityTypes.Track
+      ? notif.initiator
+      : notif.metadata.entity_id
+
+  // Action table entity is trackId for CreateTrack notifications
+  // Allowing multiple track creates to be associated w/ a single notif for your subscription
+  // For collections, the entity is the owner id, producing a distinct notif for each
+  let createdActionEntityId =
+    actionEntityType === actionEntityTypes.Track
+      ? notif.metadata.entity_id
+      : notif.metadata.entity_owner_id
+
+  // Create notification for each subscriber
+  await Promise.all(subscribers.map(async (s) => {
+    // Add notification for this user indicating the uploader has added a track
+    let notificationTarget = s.subscriberId
+
+    let unreadQuery = await models.Notification.findAll({
+      where: {
+        isViewed: false,
+        userId: notificationTarget,
+        type: createType,
+        entityId: notificationEntityId
+      },
+      transaction: tx
+    })
+
+    let notificationId = null
+    if (unreadQuery.length === 0) {
+      let createTrackNotifTx = await models.Notification.create({
+        isViewed: false,
+        isRead: false,
+        isHidden: false,
+        userId: notificationTarget,
+        type: createType,
+        entityId: notificationEntityId,
+        blocknumber,
+        timestamp
+      }, { transaction: tx })
+      notificationId = createTrackNotifTx.id
+    } else {
+      notificationId = unreadQuery[0].id
+    }
+
+    if (notificationId) {
+      // Action entity id can be one of album/playlist/track
+      let notifActionCreateTx = await models.NotificationAction.findOrCreate({
+        where: {
+          notificationId,
+          actionEntityType: actionEntityType,
+          actionEntityId: createdActionEntityId,
+          blocknumber
+        },
+        transaction: tx
+      })
+
+      // Update Notification table timestamp
+      let updatePerformed = notifActionCreateTx[1]
+      if (updatePerformed) {
+        let newNotificationTimestamp = notifActionCreateTx[0].createdAt
+        await models.Notification.update({
+          timestamp: newNotificationTimestamp
+        }, {
+          where: { id: notificationId },
+          returning: true,
+          plain: true,
+          transaction: tx
+        })
+      }
+    }
+  }))
+
+  // Dedupe album /playlist notification
+  if (createType === notificationTypes.Create.album ||
+      createType === notificationTypes.Create.playlist) {
+    let trackIdList = notif.metadata.collection_content.track_ids
+    if (trackIdList.length > 0) {
+      for (var entry of trackIdList) {
+        let trackId = entry.track
+        await models.NotificationAction.destroy({
+          where: {
+            actionEntityType: actionEntityTypes.Track,
+            actionEntityId: trackId
+          }
+        }, { transaction: tx })
+      }
+    }
+  }
+}
